move method to Util
