+++ src/app/shell/admin-tools/users/users.component.ts
@@ -5,7 +5,6 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { Select, Store } from '@ngxs/store';
 import { Observable, Subject } from 'rxjs';
 import { debounceTime, distinctUntilChanged, filter, takeUntil } from 'rxjs/operators';
-import { Constants } from 'src/app/shared/constants/constants';
 import { UserTabsUkr, UserTabsUkrReverse } from 'src/app/shared/enum/enumUA/tech-admin/users-tabs';
 import { NoResultsTitle } from 'src/app/shared/enum/no-results';
 import { ChildCards } from 'src/app/shared/models/child.model';
@@ -13,6 +12,7 @@ import { UsersTable } from 'src/app/shared/models/usersTable';
 import { GetChildren } from 'src/app/shared/store/admin.actions';
 import { AdminState } from 'src/app/shared/store/admin.state';
 import { UserState } from 'src/app/shared/store/user.state';
+import { Util } from 'src/app/shared/utils/utils';
 
 @Component({
   selector: 'app-users',
@@ -28,7 +28,6 @@ export class UsersComponent implements OnInit {
 
   readonly userRoleUkr = UserTabsUkr;
   readonly noUsers = NoResultsTitle.noUsers;
-  readonly constants: typeof Constants = Constants;
 
   filter = new FormControl('');
   filterValue: string;
@@ -59,27 +58,26 @@ export class UsersComponent implements OnInit {
       takeUntil(this.destroy$),
       filter((children: ChildCards) => !!children)
     ).subscribe((children: ChildCards) => {
-      this.children = this.updateStructureForTheTable(children.entities)
+      this.children = Util.updateStructureForTheTable(children.entities)
       this.allUsers = this.children;
     })
 
     // this.parents$.pipe(
     //   takeUntil(this.destroy$),
     //   filter((parents: Parent[]) => !!parents)
-    // ).subscribe((parents: Parent[]) => this.parents = this.updateStructureForTheTable(parents))
+    // ).subscribe((parents: Parent[]) => this.parents = Util.updateStructureForTheTable(parents))
     // TODO: for the tab 'Батьки' will implement when backend will be ready
 
     // this.parents$.pipe(
     //   takeUntil(this.destroy$),
     //   filter((parents: Parent[]) => !!parents),
     //   combineLatestWith(this.children$),
-    // ).subscribe(users => this.allUsers = this.updateStructureForTheTable(users))
+    // ).subscribe(users => this.allUsers = Util.updateStructureForTheTable(users))
     // this.store.dispatch(new GetParents());
     // TODO: for the tab 'Усі' will implement when backend will be ready    
 
     this.store.dispatch(new GetChildren());
   }
-
   
   /**
    * This method filter users according to selected tab
@@ -92,22 +90,5 @@ export class UsersComponent implements OnInit {
       { relativeTo: this.route }
     );
   }
-
-  updateStructureForTheTable(users): UsersTable[] {
-    let updatedUsers = [];
-    users.forEach((user) => {
-      updatedUsers.push({
-        id: user.id,
-        pib: `${user.lastName} ${user.firstName} ${user.middleName}`,
-        email: user.email || 'не вказано',
-        place: 'не вказано',
-        phoneNumber: user.phoneNumber ? `${this.constants.PHONE_PREFIX} ${user.phoneNumber}` : `не вказано`,
-        role: user.parentId ? 'Діти' : 'Батьки',
-        status: user.accountStatus || 'Accepted',
-      });
-    });
-    return updatedUsers;
-  }
-
 }
